Expand pytest fixtures that represent .csv and .parquet files
To ensure that the test suite covers a wider breadth of potential
model-output data scenarios, stop using the simple pyarrow fixture
as the basis for generating .csv and .parquet test files. Instead,
create a dictionary represntation of a model-output file and use
that to generate .csv and .parquet files on disk that we can
use throughout the test suite.