Add support for deploying the hubverse_transform package as an AWS lambda function
This changeset is the first step for getting this code deployable by
the hubverse-transform-model-output lambda function that exists in
the Hubverse's AWS account. This version of the code supports manually
updating the code in the lambda function by running bash script.

The next version will do the deployment via GitHub actions, as changes
are merged into the main branch.