Throw warnings when the incoming file is not valid
We've recently encountered two situations when testing the transform
lambdas that should not raise errors.
1. someone manually creates a folder in a hub's raw/ S3 folder, which triggers an S3 creation event
2. model-output folders contain an .md or .txt file that should not be included in the transform process

We still want the transform process to halt when one of the above scenarios occurs, but ideally
raising a warning instead of an error will help eliminate this noise when querying Cloudwatch
logs for errors.