First set of changes for HubConfig object (#21)
* Move specific pytest fixtures back to test_model_output.py

These fixtures have moved around a bit. But now that it looks like
all tests from the ModelOutputHandler class will be in a single file,
let's move the relevant pytest fixtures back to that file for
better clarity.

* Create skeleton HubConfig object

This will eventually be the object that returns a hub-specific
arrow schema for use in the transforms. This first version simply
reads in hub config files.

* Add file-level mypy directives

cloudpathlib's AnyPath object doesn't play well with mypy