Fix a bug when opening a parquet file on S3
When reading parquet files from S3, hubverse-transform
does an initial read to get the schema (so we can override
it if necessary). However, the read fails because it's
reading the wrong thing, and the transform process tries
to open the model-output data on the local filesystem instead
of on S3.

I opened an issue to address the lack of S3 test cases,
which resulting in this bug hitting production:
https://github.com/hubverse-org/hubverse-transform/issues/30