Update remaining Python dependencies
Dependabot opened a PR (#38) to bump numpy from 1.26.4 to 2.2.4. Rather than do
that (upgrading to numpy 2.0 is not trivial), this upgrades all Python
dependencies (Dependabot only opens 5 PRs at once). As it turns out,
the latest version of Pyarrow doesn't require numpy, so this is a better
alternative and will also lessen the initial Dependabot noise.